Moved identity logic to Identity Service.
