+++ src/identity/identity.controller.spec.ts
@@ -49,9 +49,9 @@ describe('DidController', () => {
         '@context': 'https://www.w3.org/ns/did/v1',
         'id': `did:lto:${address}`,
         'verificationMethod': [{
-          id: `identity:lto:${address}#key`,
+          id: `did:lto:${address}#key`,
           type: 'Ed25519VerificationKey2018',
-          controller: `identity:lto:${address}`,
+          controller: `did:lto:${address}`,
           publicKeyBase58: 'AVXUh6yvPG8XYqjbUgvKeEJQDQM7DggboFjtGKS8ETRG',
         }],
         'authentication': [
