+++ src/identity/identity.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { identityProviders } from './identity.providers';
 import { IdentityController } from './identity.controller';
+import { IdentityService } from './identity.service';
 import { LoggerModule } from '../logger/logger.module';
 import { ConfigModule } from '../config/config.module';
 import { StorageModule } from '../storage/storage.module';
@@ -8,12 +8,7 @@ import { StorageModule } from '../storage/storage.module';
 export const DidModuleConfig = {
   imports: [LoggerModule, ConfigModule, StorageModule],
   controllers: [IdentityController],
-  providers: [
-    ...identityProviders,
-  ],
-  exports: [
-    ...identityProviders,
-  ],
+  providers: [IdentityService],
 };
 
 @Module(DidModuleConfig)
