--- src/identity/identity.providers.ts
@@ -1,2 +1,0 @@
-export const identityProviders = [
-];
